Backend/pyproject.toml
[project]
name = "django-udemy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "django-environ>=0.12.0",
    "django-rest-framework>=0.1.0",
]

[dependency-groups]
dev = [
    "drf-yasg>=1.21.10",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-faker>=2.0.0",
    "ruff>=0.12.10",
]

[tool.ruff]
line-length = 79                  # Máximo de caracteres por linha
extend-exclude = ['migrations']   # Ignora a pasta migrations

[tool.ruff.lint]
preview = true                    # Ativa regras experimentais
select = ['I', 'F', 'E', 'W', 'PL', 'PT'] 
# - 'I'  = isort (ordenação de imports)
# - 'F'  = pyflakes (erros básicos: variáveis não usadas, etc.)
# - 'E'  = erros de estilo (pep8)
# - 'W'  = warnings de estilo
# - 'PL' = pylint (boas práticas, nomes ruins, etc.)
# - 'PT' = pytest (melhores práticas em testes)

[tool.ruff.format]
preview = true
quote-style = 'single'            # Força aspas simples
